Initial release of basic project skeleton.
This commit brings in initial project codebase. It is just simple skeleton, but it is a stable building block for implementing future new features. Following is an incomplete list of already resolved features:

* Modular web application based on Flask with the use of advanced features like view classes, blueprints, etc.
* Web interface localization.
* Installing third party libraries for web application development with Yarn (NPM modules).
* Web interface management with Grunt (cleanup, copy libraries to appropriate places, ...).
* Project documentation using Sphinx-doc.
* Application is packaged as Python wheel.
* Master project makefile for orchestrating development all tasks.

This first release will be tagged with version v0.1.0 and it provides a good starting point for developing more projects.